pyproject.toml
[project]
name = "fedmed"
version = "0.1.0"
description = "FedMorph — Morphology-Aware Federated Learning for Medical Image Segmentation"
authors = [
    { name = "LEE JIN" }
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "flwr>=1.5.0",
    "numpy>=1.24.0",
    "tqdm>=4.67.1",
    "opencv-python>=4.11.0.86",
    "monai>=1.3.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.rye]
virtual = true
managed = true
universal = true
dev-dependencies = [
    "pytest>=8.3.1",
    "mypy>=1.11.0",
    "pre-commit>=3.7.1",
    "ipykernel>=6.29.5",
    "pytest-cov>=7.0.0",
]

# [tool.hatch.metadata]
# allow-direct-references = true

# [tool.hatch.build.targets.wheel]
# packages = ["src"]

[tool.ruff]
lint.select = ["B", "E", "F", "I", "UP", "W", "C4", "SIM", "ARG", "PIE", "T20", "Q", "N"]
lint.ignore = ["B905", "B008", "B904", "F811", "E501"]
lint.fixable = ["B", "F", "I", "UP"]
lint.unfixable = []
extend-include = ["*.ipynb"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
files = "src"
exclude = "tests"
strict = true
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = "src"
addopts = ["--import-mode=importlib"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_date_format = "%Y-%m-%dT%H:%M:%S%z"
log_cli_format = "%(levelname)-8s %(asctime)s %(name)s:%(funcName)s:L%(lineno)d %(message)s"

[dependency-groups]
dev = [
    "pandas>=2.3.3",
]

[tool.rye.scripts]
"prod" = { cmd = "python src/main.py" }
"dev" = { cmd = "python src/main.py", env = { "PYTHONDEVMODE" = "1" } }
"test" = { cmd = "pytest --durations=5", env = { "PYTHONDEVMODE" = "1" } }
"test:cov" = { cmd = "pytest --cov" }
"type" = { cmd = "mypy" }
